rename: 폴더명 변경
역할에 따른 구체화로 폴더명 수정

Related to #71
